refactor(code): redefine use of trait
